Miracle 4/requirements.txt
discord.py
PyNaCl
python-dotenv
pyyaml
aiosqlite
aiofiles